relocate DownloadFinalizedData from api to sync
