+++ beacon-chain/sync/checkpoint/weak-subjectivity_test.go
@@ -1,4 +1,4 @@
-package beacon
+package checkpoint
 
 import (
 	"bytes"
@@ -11,6 +11,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/v5/api/client"
+	"github.com/prysmaticlabs/prysm/v5/api/client/beacon"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/state"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
@@ -24,19 +25,6 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/testing/util"
 	"github.com/prysmaticlabs/prysm/v5/time/slots"
 )
-
-type testRT struct {
-	rt func(*http.Request) (*http.Response, error)
-}
-
-func (rt *testRT) RoundTrip(req *http.Request) (*http.Response, error) {
-	if rt.rt != nil {
-		return rt.rt(req)
-	}
-	return nil, errors.New("RoundTripper not implemented")
-}
-
-var _ http.RoundTripper = &testRT{}
 
 func marshalToEnvelope(val interface{}) ([]byte, error) {
 	raw, err := json.Marshal(val)
@@ -61,35 +49,6 @@ func TestMarshalToEnvelope(t *testing.T) {
 	require.NoError(t, err)
 	expected := `{"data":{"version":"Prysm/v2.0.5 (linux amd64)"}}`
 	require.Equal(t, expected, string(encoded))
-}
-
-func TestFallbackVersionCheck(t *testing.T) {
-	trans := &testRT{rt: func(req *http.Request) (*http.Response, error) {
-		res := &http.Response{Request: req}
-		switch req.URL.Path {
-		case getNodeVersionPath:
-			res.StatusCode = http.StatusOK
-			b := bytes.NewBuffer(nil)
-			d := struct {
-				Version string `json:"version"`
-			}{
-				Version: "Prysm/v2.0.5 (linux amd64)",
-			}
-			encoded, err := marshalToEnvelope(d)
-			require.NoError(t, err)
-			b.Write(encoded)
-			res.Body = io.NopCloser(b)
-		case getWeakSubjectivityPath:
-			res.StatusCode = http.StatusNotFound
-		}
-		return res, nil
-	}}
-
-	c, err := NewClient("http://localhost:3500", client.WithRoundTripper(trans))
-	require.NoError(t, err)
-	ctx := context.Background()
-	_, err = ComputeWeakSubjectivityCheckpoint(ctx, c)
-	require.ErrorIs(t, err, errUnsupportedPrysmCheckpointVersion)
 }
 
 func TestFname(t *testing.T) {
@@ -160,7 +119,7 @@ func TestDownloadWeakSubjectivityCheckpoint(t *testing.T) {
 
 	wsSerialized, err := wst.MarshalSSZ()
 	require.NoError(t, err)
-	expectedWSD := WeakSubjectivityData{
+	expectedWSD := beacon.WeakSubjectivityData{
 		BlockRoot: bRoot,
 		StateRoot: wRoot,
 		Epoch:     epoch,
@@ -169,7 +128,7 @@ func TestDownloadWeakSubjectivityCheckpoint(t *testing.T) {
 	trans := &testRT{rt: func(req *http.Request) (*http.Response, error) {
 		res := &http.Response{Request: req}
 		switch req.URL.Path {
-		case getWeakSubjectivityPath:
+		case beacon.GetWeakSubjectivityPath:
 			res.StatusCode = http.StatusOK
 			cp := struct {
 				Epoch string `json:"epoch"`
@@ -188,10 +147,10 @@ func TestDownloadWeakSubjectivityCheckpoint(t *testing.T) {
 			rb, err := marshalToEnvelope(wsr)
 			require.NoError(t, err)
 			res.Body = io.NopCloser(bytes.NewBuffer(rb))
-		case renderGetStatePath(IdFromSlot(wSlot)):
+		case beacon.RenderGetStatePath(beacon.IdFromSlot(wSlot)):
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(wsSerialized))
-		case renderGetBlockPath(IdFromRoot(bRoot)):
+		case beacon.RenderGetBlockPath(beacon.IdFromRoot(bRoot)):
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(serBlock))
 		}
@@ -199,7 +158,7 @@ func TestDownloadWeakSubjectivityCheckpoint(t *testing.T) {
 		return res, nil
 	}}
 
-	c, err := NewClient("http://localhost:3500", client.WithRoundTripper(trans))
+	c, err := beacon.NewClient("http://localhost:3500", client.WithRoundTripper(trans))
 	require.NoError(t, err)
 
 	wsd, err := ComputeWeakSubjectivityCheckpoint(ctx, c)
@@ -263,7 +222,7 @@ func TestDownloadBackwardsCompatibleCombined(t *testing.T) {
 	trans := &testRT{rt: func(req *http.Request) (*http.Response, error) {
 		res := &http.Response{Request: req}
 		switch req.URL.Path {
-		case getNodeVersionPath:
+		case beacon.GetNodeVersionPath:
 			res.StatusCode = http.StatusOK
 			b := bytes.NewBuffer(nil)
 			d := struct {
@@ -275,15 +234,15 @@ func TestDownloadBackwardsCompatibleCombined(t *testing.T) {
 			require.NoError(t, err)
 			b.Write(encoded)
 			res.Body = io.NopCloser(b)
-		case getWeakSubjectivityPath:
+		case beacon.GetWeakSubjectivityPath:
 			res.StatusCode = http.StatusNotFound
-		case renderGetStatePath(IdHead):
+		case beacon.RenderGetStatePath(beacon.IdHead):
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(serialized))
-		case renderGetStatePath(IdFromSlot(wSlot)):
+		case beacon.RenderGetStatePath(beacon.IdFromSlot(wSlot)):
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(wsSerialized))
-		case renderGetBlockPath(IdFromRoot(bRoot)):
+		case beacon.RenderGetBlockPath(beacon.IdFromRoot(bRoot)):
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(serBlock))
 		}
@@ -291,7 +250,7 @@ func TestDownloadBackwardsCompatibleCombined(t *testing.T) {
 		return res, nil
 	}}
 
-	c, err := NewClient("http://localhost:3500", client.WithRoundTripper(trans))
+	c, err := beacon.NewClient("http://localhost:3500", client.WithRoundTripper(trans))
 	require.NoError(t, err)
 
 	wsPub, err := ComputeWeakSubjectivityCheckpoint(ctx, c)
@@ -308,13 +267,13 @@ func TestGetWeakSubjectivityEpochFromHead(t *testing.T) {
 	require.NoError(t, err)
 	trans := &testRT{rt: func(req *http.Request) (*http.Response, error) {
 		res := &http.Response{Request: req}
-		if req.URL.Path == renderGetStatePath(IdHead) {
+		if req.URL.Path == beacon.RenderGetStatePath(beacon.IdHead) {
 			res.StatusCode = http.StatusOK
 			res.Body = io.NopCloser(bytes.NewBuffer(serialized))
 		}
 		return res, nil
 	}}
-	c, err := NewClient("http://localhost:3500", client.WithRoundTripper(trans))
+	c, err := beacon.NewClient("http://localhost:3500", client.WithRoundTripper(trans))
 	require.NoError(t, err)
 	actualEpoch, err := getWeakSubjectivityEpochFromHead(context.Background(), c)
 	require.NoError(t, err)
@@ -385,97 +344,4 @@ func populateValidators(cfg *params.BeaconChainConfig, st state.BeaconState, val
 		return err
 	}
 	return st.SetBalances(balances)
-}
-
-func TestDownloadFinalizedData(t *testing.T) {
-	ctx := context.Background()
-	cfg := params.MainnetConfig().Copy()
-
-	// avoid the altair zone because genesis tests are easier to set up
-	epoch := cfg.AltairForkEpoch - 1
-	// set up checkpoint state, using the epoch that will be computed as the ws checkpoint state based on the head state
-	slot, err := slots.EpochStart(epoch)
-	require.NoError(t, err)
-	st, err := util.NewBeaconState()
-	require.NoError(t, err)
-	fork, err := forkForEpoch(cfg, epoch)
-	require.NoError(t, err)
-	require.NoError(t, st.SetFork(fork))
-	require.NoError(t, st.SetSlot(slot))
-
-	// set up checkpoint block
-	b, err := blocks.NewSignedBeaconBlock(util.NewBeaconBlock())
-	require.NoError(t, err)
-	b, err = blocktest.SetBlockParentRoot(b, cfg.ZeroHash)
-	require.NoError(t, err)
-	b, err = blocktest.SetBlockSlot(b, slot)
-	require.NoError(t, err)
-	b, err = blocktest.SetProposerIndex(b, 0)
-	require.NoError(t, err)
-
-	// set up state header pointing at checkpoint block - this is how the block is downloaded by root
-	header, err := b.Header()
-	require.NoError(t, err)
-	require.NoError(t, st.SetLatestBlockHeader(header.Header))
-
-	// order of operations can be confusing here:
-	// - when computing the state root, make sure block header is complete, EXCEPT the state root should be zero-value
-	// - before computing the block root (to match the request route), the block should include the state root
-	//   *computed from the state with a header that does not have a state root set yet*
-	sr, err := st.HashTreeRoot(ctx)
-	require.NoError(t, err)
-
-	b, err = blocktest.SetBlockStateRoot(b, sr)
-	require.NoError(t, err)
-	mb, err := b.MarshalSSZ()
-	require.NoError(t, err)
-	br, err := b.Block().HashTreeRoot()
-	require.NoError(t, err)
-
-	ms, err := st.MarshalSSZ()
-	require.NoError(t, err)
-
-	trans := &testRT{rt: func(req *http.Request) (*http.Response, error) {
-		res := &http.Response{Request: req}
-		switch req.URL.Path {
-		case renderGetStatePath(IdFinalized):
-			res.StatusCode = http.StatusOK
-			res.Body = io.NopCloser(bytes.NewBuffer(ms))
-		case renderGetBlockPath(IdFromSlot(b.Block().Slot())):
-			res.StatusCode = http.StatusOK
-			res.Body = io.NopCloser(bytes.NewBuffer(mb))
-		default:
-			res.StatusCode = http.StatusInternalServerError
-			res.Body = io.NopCloser(bytes.NewBufferString(""))
-		}
-
-		return res, nil
-	}}
-	c, err := NewClient("http://localhost:3500", client.WithRoundTripper(trans))
-	require.NoError(t, err)
-	// sanity check before we go through checkpoint
-	// make sure we can download the state and unmarshal it with the VersionedUnmarshaler
-	sb, err := c.GetState(ctx, IdFinalized)
-	require.NoError(t, err)
-	require.Equal(t, true, bytes.Equal(sb, ms))
-	vu, err := detect.FromState(sb)
-	require.NoError(t, err)
-	us, err := vu.UnmarshalBeaconState(sb)
-	require.NoError(t, err)
-	ushtr, err := us.HashTreeRoot(ctx)
-	require.NoError(t, err)
-	require.Equal(t, sr, ushtr)
-
-	expected := &OriginData{
-		sb: ms,
-		bb: mb,
-		br: br,
-		sr: sr,
-	}
-	od, err := DownloadFinalizedData(ctx, c)
-	require.NoError(t, err)
-	require.Equal(t, true, bytes.Equal(expected.sb, od.sb))
-	require.Equal(t, true, bytes.Equal(expected.bb, od.bb))
-	require.Equal(t, expected.br, od.br)
-	require.Equal(t, expected.sr, od.sr)
 }
