+++ changelog/kasey_refactor-checkpoint-download.md
@@ -1,0 +1,2 @@
+### Changed
+- DownloadFinalizedData has moved from the api/client package to beacon-chain/sync/checkpoint
