+++ cmd/prysmctl/weaksubjectivity/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     deps = [
         "//api/client:go_default_library",
         "//api/client/beacon:go_default_library",
+        "//beacon-chain/sync/checkpoint:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
